requirements.txt
anyio==4.6.2.post1
APScheduler==3.6.3
cachetools==4.2.2
certifi==2024.8.30
chardet==5.2.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
pillow==11.0.0
psycopg2-binary==2.9.10
PyMuPDF==1.24.13
PyPDF2==3.0.1
python-dotenv==1.0.1
python-telegram-bot==20.3
pytz==2024.2
reportlab==4.2.5
six==1.16.0
sniffio==1.3.1
tornado==6.4.1
tzdata==2024.2
tzlocal==5.2
